MainPanel 내 좌측 출력 패널을 새로 정의한 LeftSpacePanel로 변경
- WritingEditorPanel을 LeftSpacePanel의 하위 패널화
